Change used percentage values to be int
Leave the free percentage as a float64 for better precision during calculations and round to int just for the API - generate doesn't like floats and it's more readable as an int

+++ internal/metrics/source/source_test.go
@@ -18,7 +18,7 @@ var _ = Describe("Source", func() {
 				capacity           int
 				available          int
 				wantFreePercentage float64
-				wantUsedPercentage float64
+				wantUsedPercentage int
 			}{
 				{capacity: 1000, available: 100, wantFreePercentage: 10.0, wantUsedPercentage: 90.0},
 				{capacity: 1000, available: 200, wantFreePercentage: 20.0, wantUsedPercentage: 80.0},
